Loading and scrolling data from firebase

+++ app/src/main/java/com/medprimetech/plants/DataModel.java
@@ -8,18 +8,18 @@ public class DataModel {
 
     String name;
     String author;
-    int id_;
+    double id_;
     Bitmap image;
     String imageURL;
 
-    public DataModel(String name, String author, int id_, Bitmap image) {
+    public DataModel(String name, String author, double id_, Bitmap image) {
         this.name = name;
         this.author = author;
         this.id_ = id_;
         this.image=image;
     }
 
-    public DataModel(String name, String author, int id_, String imageURL) {
+    public DataModel(String name, String author, double id_, String imageURL) {
         this.name = name;
         this.author = author;
         this.id_ = id_;
@@ -40,7 +40,7 @@ public class DataModel {
         return image;
     }
 
-    public int getId() {
+    public double getId() {
         return id_;
     }
 }
